Fixed Index filters and paging + Lookup selectlist

+++ SamaraDB/dbo/Tables/Item.sql
@@ -3,7 +3,7 @@
 	[ItemID] INT NOT NULL PRIMARY KEY IDENTITY,
 	[ItemName] VARCHAR(100) NULL, 
     [UnitID] INT NULL, 
-    [ReorderLevel] INT NULL, 
+    [ReorderLevel] DECIMAL(18, 2) NULL, 
     [TaxPerc] DECIMAL(18, 2) NULL, 
     CONSTRAINT [FK_Item_Units] FOREIGN KEY ([UnitID]) REFERENCES [Units]([UnitID])
 )
